backend/pyproject.toml
 [tool.poetry]
 name = "tara-backend"
 version = "0.1.0"
 description = "FastAPI backend for the TARA clinical assistant"
 authors = ["Your Team <team@example.com>"]
 readme = "README.md"
 packages = [{ include = "app" }]

 [tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
httpx = "^0.27.2"
openai = "^1.55.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
pgvector = "^0.2.5"
python-dotenv = "^1.0.1"

 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.2"
 pytest-asyncio = "^0.24.0"

 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
